+++ tests/ClientTest.php
@@ -19,8 +19,7 @@ class ClientTest extends TestCase
 {
     private $testConfig = [
         'uri' => 'https://maxemail.example.com/',
-        'username' => 'api@user.com',
-        'password' => 'apipass',
+        'token' => 'apitoken',
     ];
 
     public function testConfigValid()
@@ -31,16 +30,11 @@ class ClientTest extends TestCase
             $expectedUri = $this->testConfig['uri'] . 'api/json/';
             static::assertSame($expectedUri, $actual['base_uri']);
 
-            $expectedAuth = [
-                $this->testConfig['username'],
-                $this->testConfig['password'],
-            ];
-            static::assertSame($expectedAuth, $actual['auth']);
-
             $expectedHeaders = [
                 'User-Agent' => 'MxmApiClient/' . Client::VERSION . ' PHP/' . PHP_VERSION,
                 'Content-Type' => 'application/x-www-form-urlencoded',
                 'Accept' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->testConfig['token'],
             ];
             static::assertSame($expectedHeaders, $actual['headers']);
 
@@ -81,8 +75,7 @@ class ClientTest extends TestCase
     public function testConfigDefaultHost()
     {
         $config = [
-            'username' => 'api@user.com',
-            'password' => 'apipass',
+            'token' => 'apitoken',
         ];
 
         $api = new Client($config);
@@ -104,8 +97,7 @@ class ClientTest extends TestCase
     {
         $config = [
             'uri' => 'https://maxemail.example.com/some/extra/path',
-            'username' => 'api@user.com',
-            'password' => 'apipass',
+            'token' => 'apitoken',
         ];
 
         $api = new Client($config);
@@ -130,8 +122,7 @@ class ClientTest extends TestCase
 
         $config = [
             'uri' => '//',
-            'username' => 'api@user.com',
-            'password' => 'apipass',
+            'token' => 'apitoken',
         ];
 
         new Client($config);
@@ -144,17 +135,49 @@ class ClientTest extends TestCase
 
         $config = [
             'uri' => 'maxemail.example.com',
-            'username' => 'api@user.com',
-            'password' => 'apipass',
+            'token' => 'apitoken',
         ];
 
         new Client($config);
     }
 
+    public function testConfigLegacyAuthentication(): void
+    {
+        $config = [
+            'username' => 'api@user.com',
+            'password' => 'apipass',
+        ];
+
+        $api = new Client($config);
+
+        $factory = function (array $actual) use ($config): GuzzleClient {
+            $expectedAuth = [
+                $config['username'],
+                $config['password'],
+            ];
+            static::assertSame($expectedAuth, $actual['auth']);
+
+            return $this->createMock(GuzzleClient::class);
+        };
+
+        $api->setHttpClientFactory($factory);
+
+        // Get a service, to trigger the HTTP Client factory
+        $api->folder;
+    }
+
+    public function testConfigMissingToken(): void
+    {
+        $this->expectException(Exception\InvalidArgumentException::class);
+        $this->expectExceptionMessage('API config requires token OR username & password');
+
+        new Client([]);
+    }
+
     public function testConfigMissingUsername(): void
     {
         $this->expectException(Exception\InvalidArgumentException::class);
-        $this->expectExceptionMessage('API config requires username & password');
+        $this->expectExceptionMessage('API config requires token OR username & password');
 
         $config = [
             'password' => 'apipass',
@@ -166,7 +189,7 @@ class ClientTest extends TestCase
     public function testConfigMissingPassword(): void
     {
         $this->expectException(Exception\InvalidArgumentException::class);
-        $this->expectExceptionMessage('API config requires username & password');
+        $this->expectExceptionMessage('API config requires token OR username & password');
 
         $config = [
             'username' => 'api@user.com',
@@ -175,11 +198,30 @@ class ClientTest extends TestCase
         new Client($config);
     }
 
-    public function testGetConfig(): void
+    public function testGetConfigWithToken(): void
     {
         $api = new Client($this->testConfig);
 
-        static::assertSame($this->testConfig, $api->getConfig());
+        $expected = [
+            'uri' => $this->testConfig['uri'],
+            'username' => null,
+            'password' => null,
+        ];
+
+        static::assertSame($expected, $api->getConfig());
+    }
+
+    public function testGetConfigWithLegacyAuthentication(): void
+    {
+        $config = [
+            'uri' => 'https://maxemail.example.com/',
+            'username' => 'api@user.com',
+            'password' => 'apipass',
+        ];
+
+        $api = new Client($config);
+
+        static::assertSame($config, $api->getConfig());
     }
 
     public function testSetGetLogger()