+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support for API token authentication. Username and password can still be used
+  as a fallback.
 ### Changed
 - Deprecate `getConfig()` method. Packages can maintain their config internally.